requirements.txt
# Core Framework
Django==5.2.8
asgiref==3.11.0

# Database
psycopg2-binary==2.9.9

# REST API
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1

# Admin Enhancements
django-nested-admin==4.0.2

# API Documentation
drf-yasg==1.21.11

# Utilities
django-filter==25.2
inflection==0.5.1
Markdown==3.10
packaging==25.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0

# Production (commented - uncomment for deployment)
# gunicorn==21.2.0
# whitenoise==6.6.0
# django-cors-headers==4.3.1
# celery==5.3.4
# redis==5.0.1
# sentry-sdk==1.39.1
